isolate CalibrationFrame and CalibrationCamera as objects
