+++ test/test_calibration/test_camera_calibration_frames.py
@@ -30,7 +30,7 @@ up->down along image height
 from __future__ import division, print_function
 import numpy
 
-from openalea.phenomenal.calibration import CalibrationCamera, CalibrationFrame
+from openalea.phenomenal.calibration.object import CalibrationFrame, CalibrationCamera
 
 
 def test_image_frame():
